Add mnemonic package
Handles converting key bytes into a string of 24 words from a
given dictionary of 1626 words, and vice versa. This is the base
mnemonic.

Mnemonics can also have a checksum word. It is a repeated word from the
base mnemonic. The word is selected as a crc32 checksum of
the concatenated utf-8 prefixes of each base word, modulo 24.